moved format func to format.go file
